Fix circular imports, add test script for circular imports (#1450)
