+++ .changeset/bright-baboons-ring.md
@@ -1,0 +1,8 @@
+---
+'houdini-plugin-svelte-global-stores': patch
+'houdini-svelte': patch
+'houdini-react': patch
+'houdini': patch
+---
+
+fix internal circular dependencies
