+++ package.json
@@ -8,6 +8,7 @@
 		"tests": "vitest",
 		"tests:ui": "vitest --ui --coverage",
 		"test": "pnpm run tests",
+		"check-circular-deps": "pnpm npx madge --circular --extensions ts,tsx ./",
 		"build:all": "turbo build",
 		"build": "turbo run build --filter=\"./packages/*\"",
 		"dev": "turbo dev --filter=\"./packages/*\"",
@@ -39,6 +40,7 @@
 		"eslint-plugin-unused-imports": "^3.0.0",
 		"graphql": "^15.8.0",
 		"lint-staged": "^12.3.4",
+		"madge": "^8.0.0",
 		"prettier": "^2.8.3",
 		"turbo": "^1.8.8",
 		"typescript": "^4.9",
