+++ packages/houdini-react/src/plugin/state.ts
@@ -1,0 +1,7 @@
+import type { ProjectManifest } from 'houdini'
+
+export let manifest: ProjectManifest
+
+export function setManifest(newManifest: ProjectManifest): void {
+	manifest = newManifest
+}
