+++ packages/houdini-react/src/runtime/routing/index.ts
@@ -1,2 +1,3 @@
 export * from './Router'
+export * from './hooks'
 export { type SuspenseCache, suspense_cache } from './cache'
