+++ packages/houdini-svelte/src/plugin/extractLoadFunction.ts
@@ -6,12 +6,16 @@ import { fs, parseJS, path } from 'houdini'
 import type * as recast from 'recast'
 import { transformWithEsbuild } from 'vite'
 
-import type { HoudiniRouteScript } from './kit'
-import { stores_directory_name, store_suffix } from './kit'
 import { houdini_load_fn } from './naming'
+import { stores_directory_name, store_suffix } from './storeConfig'
 
 type Program = recast.types.namedTypes.Program
 type VariableDeclaration = recast.types.namedTypes.VariableDeclaration
+
+export type HoudiniRouteScript = {
+	houdini_load?: graphql.OperationDefinitionNode[]
+	exports: string[]
+}
 
 export async function extract_load_function(
 	config: Config,
