+++ packages/houdini-svelte/src/plugin/state.ts
@@ -1,0 +1,7 @@
+import type { Config } from 'houdini'
+
+export let _config: Config
+
+export function setConfig(config: Config) {
+	_config = config
+}
