+++ packages/houdini-svelte/src/plugin/transforms/index.ts
@@ -4,8 +4,9 @@ import type { TransformPage } from 'houdini/vite'
 import * as recast from 'recast'
 import type { SourceMapInput } from 'rollup'
 
+import { plugin_config } from '../config'
 import { parseSvelte } from '../extract'
-import { plugin_config, type Framework } from '../kit'
+import { type Framework } from '../kit'
 import query from './componentQuery'
 import kit from './kit'
 import tags from './tags'
