+++ packages/houdini/src/lib/typescript.ts
@@ -2,8 +2,11 @@ import type { StatementKind, TSTypeKind } from 'ast-types/lib/gen/kinds'
 import * as graphql from 'graphql'
 import * as recast from 'recast'
 
-import { ensureImports, type Config, path, fs } from '.'
+import type { Config } from './config'
+import * as fs from './fs'
 import { unwrapType, TypeWrapper } from './graphql'
+import { ensureImports } from './imports'
+import * as path from './path'
 
 const AST = recast.types.builders
