+++ packages/plugin-svelte-global-stores/src/plugin/codegen/stores/fragment.ts
@@ -1,7 +1,10 @@
 import type { Document, GenerateHookInput } from 'houdini'
 import { fs, path } from 'houdini'
 
-import { store_name, stores_directory_name } from '../../../../../houdini-svelte/src/plugin/kit'
+import {
+	store_name,
+	stores_directory_name,
+} from '../../../../../houdini-svelte/src/plugin/storeConfig'
 import { global_store_name, global_stores_directory } from '../../kit'
 
 export async function fragmentStore({ config, pluginRoot }: GenerateHookInput, doc: Document) {
