+++ packages/plugin-svelte-global-stores/src/plugin/codegen/stores/subscription.ts
@@ -1,7 +1,10 @@
 import type { Document, GenerateHookInput } from 'houdini'
 import { fs, path } from 'houdini'
 
-import { stores_directory_name, store_name } from '../../../../../houdini-svelte/src/plugin/kit'
+import {
+	stores_directory_name,
+	store_name,
+} from '../../../../../houdini-svelte/src/plugin/storeConfig'
 import { global_stores_directory, global_store_name } from '../../kit'
 
 export async function subscriptionStore({ config, pluginRoot }: GenerateHookInput, doc: Document) {
